+++ libmx/src/list/mx_list_to_array.c
@@ -5,7 +5,7 @@ void **mx_list_to_array(t_list *list) {
     void **array = NULL;
 
     if (node && (array = (void **)malloc(sizeof(void *) * list->size)))
-        for (int i = 0; node; ++i, node = node->next)
+        for (size_t i = 0; node; ++i, node = node->next)
             array[i] = node->data;
     return array;
 }
